boot: Build example boot images with debug info
Assemble via NASM -> ld -> objcopy using ELF objects,
so we can keep debug information in the intermediate ELF objects,
and use them with gdb later on.

Also add a simple Makefile to automate building
both raw binaries from NASM directly, or intermediate ELF objects.

As an aside, add a "bits" directive to make it explicit
that we are assembling 16-bit code for the 8086.

Signed-off-by: Vangelis Koukis <vkoukis@gmail.com>